move SoftwareManagerInfo screen to BackupRestore.py add helper class to support menu.xml
